Zoom and developer settings

+++ src/main/java/com/noticemedan/map/model/kdtree/ForestInterface.java
@@ -7,7 +7,6 @@ import com.noticemedan.map.model.utilities.Rect;
 import java.util.List;
 
 public interface ForestInterface {
-	public List<OsmElement> rangeSearch(Rect query, int zoomLevel);
-
+	public List<OsmElement> rangeSearch(Rect query, double zoomLevel);
 	public OsmElement nearestNeighbor(double x, double y);
 }
